[ARXIVCE-2542] use new preflight and zzrm in tex2pdf, change internal wiring

+++ tex2pdf_service/tests/test_docker.py
@@ -3,15 +3,20 @@ import tempfile
 import time
 import requests
 import subprocess
+import urllib.parse
 import pytest
 from bin.compile_submissions import get_outcome_meta
 import logging
+import json
 
 PORT = 33031
 
-def submit_tarball(service: str, tarball: str, outcome_file: str, tex2pdf_timeout: int = 30, post_timeout: int = 10) -> None | dict:
+def submit_tarball(service: str, tarball: str, outcome_file: str, tex2pdf_timeout: int = 30, post_timeout: int = 10, json_response: bool = False, api_args: dict= {}) -> None | dict:
     meta = None
-    url = service + f"/?timeout={tex2pdf_timeout}"
+    params_dict = { "timeout": tex2pdf_timeout }
+    params_dict.update(api_args)
+    params = urllib.parse.urlencode(params_dict)
+    url = f"{service}/?{params}"
     with open(tarball, "rb") as data_fd:
         uploading = {'incoming': (os.path.basename(tarball), data_fd, 'application/gzip')}
         while True:
@@ -26,10 +31,12 @@ def submit_tarball(service: str, tarball: str, outcome_file: str, tex2pdf_timeou
 
                 if status_code == 200:
                     if res.content:
-                        with open(outcome_file, "wb") as out:
-                            out.write(res.content)
-                        meta, lines, clsfiles, styfiles, pdfchecksum = get_outcome_meta(
-                            outcome_file)
+                        if json_response:
+                            return json.loads(res.content)
+                        else:
+                            with open(outcome_file, "wb") as out:
+                                out.write(res.content)
+                            meta, lines, clsfiles, styfiles, pdfchecksum = get_outcome_meta(outcome_file)
                 else:
                     logging.warning(f"%s: status code %d", url, status_code)
 
@@ -43,39 +50,41 @@ def submit_tarball(service: str, tarball: str, outcome_file: str, tex2pdf_timeou
     return meta
 
 @pytest.fixture(scope="module")
-def docker_container():
+def docker_container(request):
+    global PORT
+    PORT = request.config.getoption('--docker-port')
+    url = f"http://localhost:{PORT}"
+
     os.makedirs("tests/output", exist_ok=True)
 
-    image_name = "public-tex2pdf-app-2024-2024-07-21"
-    container_name = "test-arxiv-tex2pdf"
-    dockerport = "8080"
+    if not request.config.getoption('--no-docker-setup'):
+        image_name = "public-tex2pdf-app-2024-2024-07-21"
+        container_name = "test-arxiv-tex2pdf"
+        dockerport = "8080"
 
-    subprocess.call(["docker", "kill", container_name])
+        subprocess.call(["docker", "kill", container_name])
 
-    # Make sure the container is the latest
-    args = ["make", "app.docker"]
-    make = subprocess.run(args, encoding='utf-8', stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    if make.returncode != 0:
-        print(make.stdout)
-        print(make.stderr)
-        pass
+        # Make sure the container is the latest
+        args = ["make", "app.docker"]
+        make = subprocess.run(args, encoding='utf-8', stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        if make.returncode != 0:
+            print(make.stdout)
+            print(make.stderr)
+            pass
 
-    # Start the container
-    args = ["docker", "run", '--security-opt', "no-new-privileges=true", "--cpus", "1", "--rm",
-            "-d",
-            "-p", f"{PORT}:{dockerport}",
-            "-e", f"PORT={dockerport}",
-            "--name", container_name,
-            image_name]
-    docker = subprocess.run(args, encoding='utf-8', stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    if docker.returncode != 0:
-        logging.error("tex2pdf container did not start")
-        pass
-
-    # container_id = docker.stdout
+        # Start the container
+        args = ["docker", "run", '--security-opt', "no-new-privileges=true", "--cpus", "1", "--rm",
+                "-d",
+                "-p", f"{PORT}:{dockerport}",
+                "-e", f"PORT={dockerport}",
+                "--name", container_name,
+                image_name]
+        docker = subprocess.run(args, encoding='utf-8', stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        if docker.returncode != 0:
+            logging.error("tex2pdf container did not start")
+            pass
 
     # Wait for the API to be ready
-    url = f"http://localhost:{PORT}"
     for _ in range(60):  # retries for 60 seconds
         try:
             response = requests.get(url)
@@ -89,10 +98,11 @@ def docker_container():
 
     yield url
 
-    # Stop the container after tests
-    with open("tex2pdf.log", "w", encoding="utf-8") as log:
-        subprocess.call(["docker", "logs", container_name], stdout=log, stderr=log)
-    subprocess.call(["docker", "kill", container_name])
+    if not request.config.getoption('--no-docker-setup'):
+        # Stop the container after tests
+        with open("tex2pdf.log", "w", encoding="utf-8") as log:
+            subprocess.call(["docker", "logs", container_name], stdout=log, stderr=log)
+        subprocess.call(["docker", "kill", container_name])
 
 @pytest.mark.integration
 def test_api_hello(docker_container):
@@ -111,7 +121,7 @@ def test_api_smoke(docker_container):
     url = docker_container + "/convert"
     tarball = "tests/fixture/tarballs/test1/test1.tar.gz"
     outcome = "tests/output/test1.outcome.tar.gz"
-    meta = submit_tarball(url, tarball, outcome)
+    meta = submit_tarball(url, tarball, outcome, api_args={"auto_detect": "true"})
     assert meta is not None
 
 
@@ -120,20 +130,19 @@ def test_api_test2(docker_container):
     url = docker_container + "/convert"
     tarball = "tests/fixture/tarballs/test2/test2.tar.gz"
     outcome = "tests/output/test2.outcome.tar.gz"
-    meta = submit_tarball(url, tarball, outcome)
+    meta = submit_tarball(url, tarball, outcome, api_args={"auto_detect": "true"})
     assert meta is not None
     assert meta.get("pdf_file") == "test2.pdf"
-    assert meta.get("tex_files") == ['fake-file-2.tex', 'fake-file-1.tex']
+    assert meta.get("tex_files") == ['fake-file-2.tex']
     # autotex says that the documents are combined alphabetically
-    assert meta.get("documents") == ['out/fake-file-1.pdf', 'out/fake-file-2.pdf']
-
+    assert meta.get("documents") == ['out/fake-file-2.pdf']
 
 @pytest.mark.integration
 def test_api_test3(docker_container):
     url = docker_container + "/convert"
     tarball = "tests/fixture/tarballs/test3/test3.tar.gz"
     outcome = "tests/output/test3.outcome.tar.gz"
-    meta = submit_tarball(url, tarball, outcome)
+    meta = submit_tarball(url, tarball, outcome, api_args={"auto_detect": "true"})
     assert meta is not None
     assert meta.get("pdf_file") == "test3.pdf"
     assert meta.get("tex_files") == ['fake-file-2.tex', 'fake-file-1.tex', 'fake-file-3.tex']
@@ -147,11 +156,23 @@ def test_api_test4(docker_container):
     url = docker_container + "/convert"
     tarball = "tests/fixture/tarballs/test4/test4.tar.gz"
     outcome = "tests/output/test4.outcome.tar.gz"
-    meta = submit_tarball(url, tarball, outcome)
+    meta = submit_tarball(url, tarball, outcome, api_args={"auto_detect": "true"})
     assert meta is not None
     assert meta.get("pdf_file") == "test4.pdf"
     assert meta.get("tex_files") == ['main.tex', 'gdp.tex']
-    # There is no reasons given for designating latex
-    assert meta.get("reasons") == []
     assert len(meta.get("converters", [])) == 2
     assert len(meta["converters"][0]["runs"]) == 4  # latex, latex, dvi2ps, ps2pdf
+
+@pytest.mark.integration
+def test_api_preflight(docker_container):
+    url = docker_container + "/convert"
+    tarball = "tests/fixture/tarballs/test3/test3.tar.gz"
+    outcome = "tests/output/test3.outcome.tar.gz"
+    meta = submit_tarball(url, tarball, outcome, json_response=True, api_args={"preflight": "v2"})
+    assert meta is not None
+    print(meta)
+    assert meta.get("status").get("key") == "success"
+    assert len(meta.get("detected_toplevel_files")) == 3
+    assert [f["filename"] for f in meta.get("detected_toplevel_files")] == ['fake-file-1.tex', 'fake-file-2.tex', 'fake-file-3.tex']
+
+
